+++ webextension/add-on/lib/api.js
@@ -1,0 +1,66 @@
+/**
+ * @file API communications.
+ * @author CheckMyHTTPS's team
+ * @license GPL-3.0
+ */
+
+CMH.api = {}
+
+/**
+ * @name requestFromUrl
+ * @function
+ * @param {string} urlTested - URL to check
+ * @returns {object} - certificates of resquested host and check server
+ * Request certificate info of an URL.
+ */
+CMH.api.requestFromUrl = async (urlTested) => {
+  const { host, port } = CMH.common.parseURL(urlTested)
+
+  const { cert, data:response_data, response } = await CMH.certificatesManager.getCertUrl(CMH.options.settings.checkServerUrl+'api.php?host='+encodeURIComponent(host)+'&port='+port)
+
+  // SSL Pinning
+  // if ((cert === null) || (!CMH.certificatesChecker.compareCertificateFingerprints(cert, { fingerprints: { sha1: CMH.options.settings.checkServerFingerprintsSha1, sha256: CMH.options.settings.checkServerFingerprintsSha256 } }))) {
+  //     return { error: 'SSL' }
+  // }
+
+  if (!response.ok) {
+    return { error: response_data }
+  }
+
+  return { data: response_data, cert: cert }
+}
+
+/**
+ * @name checkCheckServerApi
+ * @function
+ * @param {object} checkServer - Check server to check
+ * @returns {boolean}
+ * Check if a check server API is valid
+ */
+CMH.api.checkCheckServerApi = async (checkServer) => {
+  try {
+    const url = new URL(checkServer.server)
+  } catch (e) {
+    if (e instanceof TypeError) {
+      return false
+    }
+  }
+
+  const { host:defaultCheckServerHost, port:defaultCheckServerPort } = CMH.common.parseURL(CMH.options.defaultCheckServer.url)
+  const { cert, data:response_data, response } = await CMH.certificatesManager.getCertUrl(checkServer.server+'api.php?host='+encodeURIComponent(defaultCheckServerHost)+'&port='+defaultCheckServerPort)
+
+  if (!response.ok) {
+    CMH.ui.notification.show(_('l_serverUnreachable'))
+    return false
+  }
+
+  if ((cert === null) || (!CMH.certificatesChecker.compareCertificateFingerprints(cert, { fingerprints: { sha1: checkServer.sha1, sha256: checkServer.sha256 } }))) {
+    return false
+  }
+
+  if ((response_data === null) || (!CMH.certificatesChecker.compareCertificateFingerprints(response_data, CMH.options.defaultCheckServer))) {
+    return false
+  }
+
+  return true
+}