+++ webextension/add-on/lib/common.js
@@ -44,7 +44,18 @@ CMH.common.statusCode = [
 CMH.common.parseURL = (urlStr) => {
   const url = new URL(urlStr)
   const host = url.hostname
-  const port = url.port || 443
+  let   port = url.port
+
+  if (port == '') {
+    const protocol = url.protocol.slice(0, -1)
+    if (protocol === 'http') {
+      port = 80
+    } else if (protocol === 'ftp') {
+      port = 21
+    } else {
+      port = 443
+    }
+  }
 
   return { host: host, port: port }
 }
@@ -85,4 +96,44 @@ CMH.common.compareVersion = function (versionA, versionB) {
   }
 
   return 0
+}
+
+/**
+ * @name isWebExtTlsApiSupported
+ * @function
+ * @returns {boolean} - TLS API supported or not
+ * Check if WebExtension TLS API is supported.
+ */
+CMH.common.isWebExtTlsApiSupported = () => {
+  return (typeof browser.webRequest.getSecurityInfo !== 'undefined')
+}
+
+/**
+ * @name platform
+ * @type {string}
+ * Current platform
+ */
+CMH.common.platform = undefined
+if (typeof browser.runtime.getBrowserInfo !== 'undefined') {
+  browser.runtime.getBrowserInfo().then((details) => {
+    if ((details.vendor === 'Mozilla') && (details.name === 'Fennec')) {
+      CMH.common.platform = 'mobile'
+    } else {
+      CMH.common.platform = 'desktop'
+    }
+  })
+} else {
+  CMH.common.platform = 'desktop'
+}
+
+/**
+ * @name isDesktopPlatform
+ * @function
+ * @returns {boolean} - Is desktop platform
+ * Check if the current platform is desktop.
+ */
+CMH.common.isDesktopPlatform = () => {
+  if (typeof CMH.common.platform !== 'undefined') {
+    return (CMH.common.platform === 'desktop')
+  }
 }